Allow running integration tests with JDK 21
The Kotlin compiler cannot yet produce Java 21 bytecode,
but we possibly need to set a toolchain to the Kotlin
compiler for when the target test JDK is different from
the current JDK (used as default toolchain).
Instead of always using the target test JDK as compilation
toolchain (which would break when testing with JDK 21),
then use the lowest of the two: this ensures the bytecode
will be compatible with the target test JDK, without
requiring a too recent, and Kotlin-unsupported, compilation
toolchain.